self-service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.10.6
httpx==0.28.1
pytest==8.3.4
python-dotenv==1.0.1
